requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2

aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
apispec==6.6.1
argcomplete==3.5.0
asgiref==3.8.1
async-timeout==4.0.3
asynch==0.2.5
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
black==24.10.0
blinker==1.8.2
bokeh==3.6.1
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
ciso8601==2.3.2
click==8.1.7
clickclick==20.10.2
clickhouse-driver==0.2.9
# clickhouse-sqlalchemy==0.3.2
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.8.2
configupdater==3.2
confluent-kafka==2.6.0
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==43.0.1
dask==2024.11.1
dask-expr==1.1.18
deprecated==1.2.14
dill==0.3.8
distributed==2024.11.1
dnspython==2.6.1
docutils==0.21.2
email-validator==2.2.0
exceptiongroup==1.2.2
frozenlist==1.4.1
fsspec==2024.9.0
geographiclib==2.0
geopy==2.4.1
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
lazy-object-proxy==1.10.0
loguru==0.7.2
lz4==4.3.3
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.22.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
opentelemetry-api==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
ordered-set==4.1.0
packaging==24.1
pandas==2.1.4
partd==1.4.2
pathspec==0.12.1
pendulum==3.0.0
pika==1.3.2
pillow==11.0.0
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==18.1.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.1
pydantic-core==2.23.3
pydantic-settings==2.6.1
pygments==2.18.0
pyjwt==2.9.0
pyspark==3.5.3
pytz==2024.2
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.38
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.2
sqlparse==0.5.1
tabulate==0.9.0
tblib==3.0.0
tenacity==9.0.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2024.1
tzlocal==5.3
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.2.5
urllib3==2.2.3
werkzeug==2.2.3
wheel==0.44.0

